WIP: Introduce BuildConfigNoFile
I'm not convinced it's worth this massive overhaul. This expresses, at
the type level, that builds can either have config files or not. This
can allow us to guarantee things like not building local packages
without config files, which is the desired behavior for the script
command.

Downside: this is a huge amount of code change for a relatively minor
corner case. On the other hand, perhaps we will, in the future, want to
give more general access to no-config-file functionality. For example, I
can picture it being useful for the `exec` family of commands.